+++ src/backend/Services/Task/TaskService.Infrastructure/Extensions/ServiceExtensions.cs
@@ -1,11 +1,11 @@
-﻿using System.Runtime.CompilerServices;
-using Microsoft.EntityFrameworkCore;
+﻿using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using TaskService.Infrastructure.Extensions.Cache;
 using TaskService.Infrastructure.Extensions.Services.FileStorage;
 using TaskService.Infrastructure.Interceptors;
 using TaskService.Infrastructure.Persistence;
+using TaskService.Infrastructure.Services;
 
 namespace TaskService.Infrastructure.Extensions;
 
@@ -15,28 +15,31 @@ public static class ServiceExtensions
     {
         public void ConfigureInfrastructureLayer(IConfiguration configuration)
         {
+            services.AddScoped<IssueNotificationService>();
+
             services.AddSingleton<SoftDeleteInterceptor>();
             services.AddDbContext<TaskServiceDbContext>((sp, options) =>
+            {
+                options.UseNpgsql(configuration.GetConnectionString("DefaultConnection"));
+                options.AddInterceptors(sp.GetRequiredService<SoftDeleteInterceptor>());
+                options.UseAsyncSeeding(async (context, _, cancellationToken) =>
                 {
-                    options.UseNpgsql(configuration.GetConnectionString("DefaultConnection"));
-                    options.AddInterceptors(sp.GetRequiredService<SoftDeleteInterceptor>());
-                    options.UseAsyncSeeding(async (context, _, cancellationToken) =>
-                    {
-                        FakeDataFactory fake = new FakeDataFactory();
-                        await fake.Seed((TaskServiceDbContext)context, cancellationToken);
-                    });
-                    options.UseSeeding((context, _) =>
-                    {
-                        FakeDataFactory fake = new FakeDataFactory();
-                        fake.Seed((TaskServiceDbContext)context);
-                    });
+                    var fake = new FakeDataFactory();
+                    await fake.Seed((TaskServiceDbContext)context, cancellationToken);
                 });
+                options.UseSeeding((context, _) =>
+                {
+                    var fake = new FakeDataFactory();
+                    fake.Seed((TaskServiceDbContext)context);
+                });
+            });
 
 
             services.AddCache(configuration);
             services.ConfigureGrpcFileStorageClient(configuration);
         }
     }
+
     extension(IServiceProvider provider)
     {
         public async Task InitializeDatabaseAsync()